stderr output capture for user facing interface functions
Added a decorator to the functions in UniversalInterface
which are meant to be called by the user. Since these are
the only entry points into underlying packages, these will
guarantee excess (and often rubish) warnings from the
3rd party packages are not dumped straight back to the user.

Unit tests for each of the entry points are included, making
use of a mock interface object.