Improvements to data method toString()
If an axis has all default names, they are omitted in the
output, replaced with an empty column or row instead. The
placeholder for omitted columns is now '--' instead of
'...' since the later was also used to indicate truncated
point or feature names. toString and show now allow control
of the lenght of point or feature names printed.