Rework of scikitlearn's getAttributes.
Implemented a helper which will collect and screen possible model
attributes of learner which is a python object. The helper is general
in that it is controlled by several input functions. This was then deployed
to the scikitlearn interface, and demonstrated non-crashing funcitonality
with a unit test.

Universal interface itself was modified to ensure that input parameters
are always included in the output of getAttributes.