Implemented top level function trainAndTestOnTrainingData()
This is the function to be used for learning in that case that no
testing data has been withheld. By default, we will return the training
error of the model. However, the flag 'crossValidationError' if set to
True, will cause this function to partition the data into folds, train
and test a learner for cross validation fold, and return a value which
is the average of the results for each fold. Consequently, using
this function means there cannot be any free parameters provided by
the user.