Added plotting methods to data objects
Includes three kinds of plotting: plot() for a heatmap of the
whole object, point and feature distrubutions for a histogram
of the values in a specific point or feature, and point and
feature comparison, for the pairwise scatterplot of the
inputs. Commit includes an example script displaying these
new methods.