Improved safety of automatic learner registration
When UML is loading, the configuration file is read
in, and those entries in the section RegisteredLearners
are attempted to be registered as custom learners.
This process is now safe, in that if importing
the learner objects fail, no exception is raised,
only a warning is printed to standard error.