Reworked the implementation of changes in SessionConfiguration
Prior to this commit, temporary changes and saving those changes
operated correctly, but had a two stage implementation that obscured
the history of changes. Now, instead of storing some changes in
an extra dictionary, and others in the in-memory portion of the
ConfigParser, the dictory holds all changes, and the ConfigParser
reflects what is on the disk. The difference is therefore easily
queriable if needed.