Added file output to data plotting methods
Each of the data plotting methods now takes an outPath
argument, which is None by default. When None, the plotting
methods all output to the screen; if non-None, then it is taken
to be a path specifying the desired location of the output
file. The output format is attempted to be taken from the
path's extension; failing that, a png image is default.

In order to unit test this feature, private, blocking versions
of all of the helper functions were created, but the user
facing functionality remains the same.

The plotting example now has the capacity to output plots
to file, if a directory path is provided when the script is called.