Added BaseView, MatrixView. Made associated corrections.
Added two new objects to the data hierarchy. BaseView (a subclass
of Base) disallows methods which modify the object, and acts
as a passthrough for names and paths of a concrete object. MatrixView
is a subclass first of BaseView and then Matrix, inherits the
read-only-ness of BaseView along with the implementations of
other methods of Matrix, resulting in a Matrix-like read only
window into the data of a different Matrix object.

Corrections to other parts of the data hierarchy needed to be
made to enforce object oriented access of attributes of the data.
Otherwise, there the inherited methods of MatrixView could
not function. MatrixView is included in the data test
suite, where changes enforcing object oriented access were
also made.