Point/Feature selection parameters added to createData +tests
These parameters allow you to use a list to specify indices or names
of the points and features you want to be in the returned object.
By default, a flag 'all' indicates to use all of the raw data available.
Current implementation operates on in-memory raw data only, so the
parameters cannot be used to limit how much of an input file is
stored (an improved implementation will be forthcoming). Unit test
confirm the correct ouput for all combinations of input and output
formats, as well as interaction with the ignoreNonNumericalFeatures
flag.