'Identifiers' for data objects may now be numpy integers
We define an identifier as something which specifies a single
point or feature in an object, either by name or by index.
These are generally validated by data objects and their types
are checked. Previously only a basestring or int (python int)
was allowed. Now numpy.interger types are also allowed.