Adjust cross validation implementation for accuracy
Previously, the results of each fold would always be evaluated
and then averaged across all folds. For some evaluation metrics,
this would result in an incorrect value. Now, the metric is
queried for an attribute indicating if this is an acceptable
strategy, and if not, the results of all folds are saved,
combined, and then evaluated as a whole.