Add scripts relating to other gender continuum experiments
These scripts are placed in their own folder in the examples module.
One relates to predicting gender from the sparse Harvard personality
data set, the other for visualizing the categorized question survey
run on guidedtrak. This is a baseline commit of established work -
but will see further expansion and refactoring