Remove some duplicate code in data/tests/structure_backend.py by combining tests across objects
The copy, retain, delete, and extract methods share the same api, and therefore the same
exceptions. Even though tests for copy are in a different test object (because it is non-
data modifying) the content for exception testing is exactly the same. A new 'shared' test
object was created for tests which can be used across both of the previously present
test objects.