Remove non-project scripts from the examples folder.
Remaining are those scripts meant to demonstrate some functionality of UML to
a new user. What has moved are scripts by us, for some actual purpose, including
old scripts from Spencer, and scripts relating to gender prediction and visualization.
These are now archived in Dropbox and not part of the project repo.