Rework the handling of labels in DataFrame so they always match position.
Previously, we were readjusting them to the default positional labels
at the point of need, instead, we now maintain that behavior at the point of
where the labels are changed.

Supporting this is the addition of a check in validation to make sure
the labels are always in a defaul state; the removal of the old helper
to do reassignment (we now do that directly); the unification of
of assignment to always use the default, efficient, RangeIndex object;
and some changes so the labels are not used/changed by other helpers.