Undo addition of sort parameter to pandas.concat calls
It's unclear if the default behavior is problematic (after removal
no tests fail), and even if so, the parameter was introduced in 0.23,
whereas we theoretically allow for 0.20 and above. Until it produces
a noticable problem, we won't makes use of the parameter, even though
in the future, the behavior will change to automatically not sort.