Correct DataFrameAxis transform to handle nimble object returns
For some pandas versions there are errors relating to the type and shape
of assignment via iloc, which was guaranteed to occur due to how
some other methods were using transform as a backend. Now, if a
nimble object return value is encountered, it is coppied to a
flattened numpy array.