Validate List's sublist type, fix related bug in points.copy
Now, all non-view Lists will check that each value in the self._data list is
also a list. The points.copy implementation wasn't correctly copying subvalues,
so a points copy of a ListView contained FeatureViewers. The implementation
has been corrected so that FeatureViewers are not placed in List objects, and
List objects don't share sublist references when doing point copies.