Autoimpute dependency wrangling
New versions of numpy have changed how they present some configuration
info, yet theano (being a dying project) has not updated to match.
Eventually pymc3 will start relying on their own fork of theano,
but currently, on numpy 1.22 and greater, we have to patch in
the information to where it is expected.