Harden np patch to resolve pymc3/theano attribute access error
Now make use of system_info to get a general replacement (or
allowed empty value)  via a fault safe method. Note: there
are potentially even more robust ways of selecting the correct
config info to patch in, but this suffices to avoid the error
(even if there is a chance theano won't be able to access a
blas package).