cluster generation changes +tests
Modified the output of generateClusteredPoints to be more
easily separable, in preparation for usage in Keras tests.
Instead of adding each cluster further out in the same dirction,
the different clusters are packed into the available feature
dimensions, one in the positive side of the axis, and one on
the negative side. If more features are being generated than
are needed, the pattern simply repeats.

A test was added checking against handmade output, and one
modified to do a parallel calcuation of the same results.