Numerous semi-related improvements to the keras interface +tests
The usage of generators for fitting and predicting have been
removed (no longer needed given our minimum version). Instead,
our Sparse data objects are converted to csr before being
passed to keras code. This
greatly simplified some of the code. The tests related to
this functionality have been removed or modified.

The parameter/defaults getters were affected by this change
as well. They were simultaneously updated to respect the
fact that the Model compile, fit, and predict API is
standard across all subclasses, and to remove uneeded ignore
values.

The tests for keras were extended to actually check against
hard values or parallel construction; not just confirm that
execution was completed without any exceptions.

This does NOT address the issue of variable handling of the
the classification output.