Modify how keras unpacks classifier output +tests
Uses the shape of the output of the network along with information
from the loss function to convert to labels in all cases of apply.

Tests check (almost) all of the various binary and multiclass loss
functions, and associated network structures. The exception is
KLDivergence, which will be handled in a subsequent PR.

Related fixes include correcting classifier learning type identification,
simplifying parameter grabing by using Model, making sure the loss
is accessible through the TrainedLearner, and  a test adjustment involving
more stringent validation of "steps" in earlier versions of keras.