Change Supervised Learning example to have correct model
The GradientBoostingRegressor ensemble estimator was only
meant for data sets of less than 10k points, and in the
1.1.0 SKL version code modifications made the efficiency
for larger datasets particularly bad. The example now uses
the HistGradientBoostingRegressor as suggested by the user
guide, which allows for reasonable runtimes regardless
of version.

This changed some of the outputs, so not only did the unit
test regex comparison need to be changed, but some of the
explanatory text no longer matched the displayed behavior.
Other small edits to the text were also added.