Support loadTrainedLearner in keras
There is now the abstract _loadTrainedLearnerBackend method
implemented for all interfaces, though other than keras it
simply raises an exception. For keras, additional changes
were needed to suport this addition. Input and output decoders
and now employed in I/O transfermation if needed. Parameter
grabbing for the keras.applications init-ers needed to be
hard-coded in some cases. A speed improvement was made
to learnerType by avoiding getAttributes.

Furthermore, in an all-interfaces change, TrainedLearner
may now a _trainXShape as None if there are no requirements
to match a training data set (as is the case with keras apps).

A unit test in in place which loads each supported pretrained
learner and verifies it can correctly predict a handful of
imagenet validation images.