Added a IPython/Notebook specific default for max width of printing.
This affects show, __repr__, and toString. We detect if we're executing
in IPython by seeing if there is a currently available interactive shell,
and assign the new default if maxHeight='automatic'. Default choosen to
be sensible even on a 1360 by 762 laptop screen.

The docstring for show has been adjusted to include this change.

The tests and outputs for the landing page examples have been adjusted
to match. Within the tests, the scripts are run in CPython, not as a
notebook as a user would, so the execution environment is adjusted to
have a terminal size to match the new default.