Re-support for local testing post build modernization
This centered around getting access to dependency information within
nimble, now that it is placed in pyproject.toml instead of a local
file (which has had its dep specifications removed). Instead, nimble
will grab the known dependency information via importlib.metadata,
or in the specific case of local testing, through a in-nimble folder
copy of pyproject.toml. As this is not the normal location for the
file, pytest had to be configured to copy it to that location at
the start of a test session.

This caused a cascade of changes relating to how pytest handles
session and per-test setup. conftest.py was required to import
nimble, but previously the nimble instantiation process required
grabbing dependency information, which meant a restructuring
so that pytest could import nimble, do test collection, prep the
pyproject.toml copy, then access nimble code reliant on dependency
information.

In practice, this means that nimble sets up the location of how
it will internally access dependencies, but that action isn't
executed until the information is needed. Thus nimble can be
instantiated before conftest.py sets up that access. Furthermore,
several test files had to be adjusted so that test collection
didn't trigger that access either.