Keras post 2.10 initializer reproducibility
Extends the keras interface so that it will construct
parameter subobjects in such a way that the seed
for that learner call can be passed in.

_inputTransformation was changed for all interfaces
so that it is aware of the learner call's seed. In
keras, that is then passed down into a specialized
helper that looks through the potential sub-objects
of each provided argument to see if it is refering
to something that takes a seed, and if so, provides
it.