Corrects and extends sparse handling in testMatch.
Now as many tests as possible still check against Sparse data.

* Previous handling of sparseUnsafeMatches reversed the desired
membership check so that those elements NOT in "sparseUnsafe" were
marked for includeSparse=False. And since the only things in
the list were single matching functions (which weren't related
to the failing tests) and tests that Sparse could pass, there
were no errors.

* The problematic np.array(data, dtype=np.object_) conversion
was deemed necessary for only List objects; limiting it to List
only removed many of the original errors.

* any/all Values_List, like allNonZero didn't need to contain
string values, and was modified.

* Building off the "sparseUnsafe" framework, the remaining tests
problematic for Sparse were categorized according to what the
results would be IF called on a valid Sparse object. They were then
skipped in the loop, and checked with constructed data
after the main loop.

* The any/all Values_str tests were deemed strictly inappropriate
for sparse. However, since any/allValues is a factory function,
it couldn't be named directly. Instead the data for any/allValues_int,
str, etc. are checked to see if it's the str one, and avoided as
needed.