Add PortableNamedTempFileContext to be used a a context manager
This is to be used in place of tempfile.NamedTemporaryFile.
It guarantees that closing will not delete the file, but that
once the context wrapper is exited, the file will always be
deleted.

This is to help ease the difference between nix and Windows
systems. In the former one can open an already open file,
in the latter, another connection can seemingly only be
opened if the delete flag is False - which causes cleanup
to become the responsibility of the caller. This commit allows
tests to be written for both system types with a minimum of pain,
requiring only that the new context manager is used, without
having to specify the flag, or handle the cleanup. In some
cases, windows may still require an explicity close operation;
this code can manage that, even if we do not currently have an
example of it