Script to be used to finalize the wheel download table in install.rst
This table is contingent on exactly which wheels were built by the
automated process, and programatically generates the table organization
appropriately.

This also rearranged install.rst so that this table could simply
be appended to the end of the file. This helps with the clarity
overall, but also hides the table in the case that it starts getting
really large.