Corrections to tests relating to decorator status for new methods / tests
The added stats methods are listed as unlogged in testLoggingCount, and the tests are
decorated accordingly. They are also, for Base listed as being
multi-dimensional allowed, while for points and features they are already
decorated as limited to 2d.

The count method was removed from Base as per previous discussion.